+++ operators/constellation-node-operator/internal/gcp/client/client_test.go
@@ -8,6 +8,7 @@ package client
 
 import (
 	"context"
+	"fmt"
 
 	compute "cloud.google.com/go/compute/apiv1"
 	"github.com/googleapis/gax-go/v2"
@@ -84,22 +85,12 @@ func (a stubInstanceTemplateAPI) Insert(ctx context.Context, req *computepb.Inse
 }
 
 type stubInstanceGroupManagersAPI struct {
-	instanceGroupManager   *computepb.InstanceGroupManager
-	getErr                 error
-	aggregatedListErr      error
-	setInstanceTemplateErr error
-	createInstancesErr     error
-	deleteInstancesErr     error
+	instanceGroupManager *computepb.InstanceGroupManager
+	aggregatedListErr    error
 }
 
 func (a stubInstanceGroupManagersAPI) Close() error {
 	return nil
-}
-
-func (a stubInstanceGroupManagersAPI) Get(ctx context.Context, req *computepb.GetInstanceGroupManagerRequest,
-	opts ...gax.CallOption,
-) (*computepb.InstanceGroupManager, error) {
-	return a.instanceGroupManager, a.getErr
 }
 
 func (a stubInstanceGroupManagersAPI) AggregatedList(ctx context.Context, req *computepb.AggregatedListInstanceGroupManagersRequest,
@@ -120,7 +111,36 @@ func (a stubInstanceGroupManagersAPI) AggregatedList(ctx context.Context, req *c
 	}
 }
 
-func (a stubInstanceGroupManagersAPI) SetInstanceTemplate(ctx context.Context, req *computepb.SetInstanceTemplateInstanceGroupManagerRequest,
+type stubRegionInstanceGroupManagersAPI struct {
+	instanceGroupManager    *computepb.InstanceGroupManager
+	managedInstance         *computepb.ManagedInstance
+	getErr                  error
+	listManagedInstancesErr error
+	setInstanceTemplateErr  error
+	createInstancesErr      error
+	deleteInstancesErr      error
+}
+
+func (a *stubRegionInstanceGroupManagersAPI) Close() error {
+	return nil
+}
+
+func (a *stubRegionInstanceGroupManagersAPI) Get(ctx context.Context, req *computepb.GetRegionInstanceGroupManagerRequest,
+	opts ...gax.CallOption,
+) (*computepb.InstanceGroupManager, error) {
+	return a.instanceGroupManager, a.getErr
+}
+
+func (a *stubRegionInstanceGroupManagersAPI) ListManagedInstances(ctx context.Context, req *computepb.ListManagedInstancesRegionInstanceGroupManagersRequest,
+	opts ...gax.CallOption,
+) ManagedInstanceIterator {
+	return &stubManagedInstanceIterator{
+		instances: []*computepb.ManagedInstance{a.managedInstance},
+		nextErr:   a.listManagedInstancesErr,
+	}
+}
+
+func (a *stubRegionInstanceGroupManagersAPI) SetInstanceTemplate(ctx context.Context, req *computepb.SetInstanceTemplateRegionInstanceGroupManagerRequest,
 	opts ...gax.CallOption,
 ) (Operation, error) {
 	return &stubOperation{
@@ -130,9 +150,15 @@ func (a stubInstanceGroupManagersAPI) SetInstanceTemplate(ctx context.Context, r
 	}, a.setInstanceTemplateErr
 }
 
-func (a stubInstanceGroupManagersAPI) CreateInstances(ctx context.Context, req *computepb.CreateInstancesInstanceGroupManagerRequest,
+func (a *stubRegionInstanceGroupManagersAPI) CreateInstances(ctx context.Context, req *computepb.CreateInstancesRegionInstanceGroupManagerRequest,
 	opts ...gax.CallOption,
 ) (Operation, error) {
+	if len(req.RegionInstanceGroupManagersCreateInstancesRequestResource.Instances) != 0 {
+		name := *req.RegionInstanceGroupManagersCreateInstancesRequestResource.Instances[0].Name
+		a.managedInstance = &computepb.ManagedInstance{
+			Instance: proto.String(fmt.Sprintf("https://www.googleapis.com/compute/v1/projects/project/zones/zone/instances/%s", name)),
+		}
+	}
 	return &stubOperation{
 		&computepb.Operation{
 			Name: proto.String("name"),
@@ -140,7 +166,7 @@ func (a stubInstanceGroupManagersAPI) CreateInstances(ctx context.Context, req *
 	}, a.createInstancesErr
 }
 
-func (a stubInstanceGroupManagersAPI) DeleteInstances(ctx context.Context, req *computepb.DeleteInstancesInstanceGroupManagerRequest,
+func (a *stubRegionInstanceGroupManagersAPI) DeleteInstances(ctx context.Context, req *computepb.DeleteInstancesRegionInstanceGroupManagerRequest,
 	opts ...gax.CallOption,
 ) (Operation, error) {
 	if a.deleteInstancesErr != nil {
@@ -201,4 +227,23 @@ func (i *stubInstanceGroupManagerScopedListIterator) Next() (compute.InstanceGro
 	pair := i.pairs[i.internalCounter]
 	i.internalCounter++
 	return pair, nil
+}
+
+type stubManagedInstanceIterator struct {
+	instances []*computepb.ManagedInstance
+	nextErr   error
+
+	internalCounter int
+}
+
+func (i *stubManagedInstanceIterator) Next() (*computepb.ManagedInstance, error) {
+	if i.nextErr != nil {
+		return nil, i.nextErr
+	}
+	if i.internalCounter >= len(i.instances) {
+		return nil, iterator.Done
+	}
+	instance := i.instances[i.internalCounter]
+	i.internalCounter++
+	return instance, nil
 }